move utils to neu.tts
