+++ speech-synthesis/TTS/tacotron/dataset.py
@@ -21,8 +21,8 @@ from nnabla.utils.data_source import DataSource
 import numpy as np
 from tqdm import tqdm
 
-from utils import audio
-from utils.text import text_normalize
+from neu.tts import audio
+from neu.tts.text import text_normalize
 
 
 class LJSpeechDataSource(DataSource):
