+++ speech-synthesis/TTS/tacotron/hparams.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from utils.hparams import HParams
+from neu.tts.hparams import HParams
 
 hparams = HParams(
